swap Double for BigDecimal for prices; fix tests

+++ src/main/java/domain/PriceList.java
@@ -5,13 +5,14 @@ import domain.constant.ServiceName;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Getter
 @AllArgsConstructor
 public class PriceList {
 
-    private final Map<ServiceName, Double> servicePrices;
-    private final Map<PackageName, Double> packagePrices;
+    private final Map<ServiceName, BigDecimal> servicePrices;
+    private final Map<PackageName, BigDecimal> packagePrices;
 
 }
